removed callbacks, renamed some classes
